backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.0,<6.0
django-cors-headers>=4.0,<5.0
mysqlclient>=2.1,<3.0
redis>=4.0,<6.0
celery>=5.0,<6.0
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
python-decouple>=3.0,<4.0
dj-database-url>=2.0,<3.0
Pillow>=9.0,<11.0
hypothesis>=6.0,<7.0
pytest>=7.0,<8.0
pytest-django>=4.0,<5.0
factory-boy>=3.0,<4.0
requests>=2.31,<3.0

# Production dependencies
gunicorn>=21.0,<22.0
django-redis>=5.0,<6.0
whitenoise>=6.0,<7.0
sentry-sdk>=1.0,<2.0
psutil>=5.9,<6.0